requirements.txt
asgiref==3.6.0
dj-database-url==1.2.0
Django==4.1.5
django-ckeditor==6.5.1
django-cockroachdb==4.1.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
gunicorn==20.1.0
Pillow==9.4.0
psycopg2==2.9.5
psycopg2-binary==2.9.5
python-dotenv==0.21.1
sqlparse==0.4.3
tzdata==2022.7
whitenoise==6.3.0
